git-svn-id: svn://192.168.1.10/public/Emutil@14716 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/emutil/D64Compare.java
@@ -247,7 +247,12 @@ public class D64Compare
 			ret.append(toHex(buf[b]));
 		return ret.toString();
 	}
-	public static short fromHex(String hex){ return ((Short)ANTI_HEX.get(hex)).shortValue();}
+
+	public static short fromHex(String hex)
+	{
+		return (ANTI_HEX.get(hex)).shortValue();
+	}
+
 	public static byte[] getFileContent(byte[][][] tsmap, int t, int mt, int s)
 	{
 		HashSet<byte[]> doneBefore=new HashSet<byte[]>();
